move more things to be under Game namespace
